requirements.txt
fastapi
uvicorn
playwright
httpx
pydantic
pyautogui
mss
pillow
